Report Creator/requirements.txt
pandas==2.1.0
pyyaml==6.0.1